refactor: 디렉토리 구조 변경
컴포넌트마다 하나의 디렉토리에 위치하도록 변경
